# optimize has_tabs_or_newline for NEON