--- a/src/unicode.cpp
+++ b/src/unicode.cpp
@@ -72,7 +72,7 @@ ada_really_inline bool has_tabs_or_newline(
    * credit for algorithmic idea: @aqrit, credit for description:
    * @DenisYaroshevskiy
    */
-  static uint8_t rnt_array[16] = {1, 0,  0, 0, 0,  0, 0, 0,
+  static uint8_t rnt_array[16] = {1, 0, 0,  0, 0, 0,  0, 0,
                                   0, 9, 10, 0, 0, 13, 0, 0};
   const uint8x16_t rnt = vld1q_u8(rnt_array);
   // m['0xd', '0xa', '0x9']