@anonrig 2024-05-01
@lemire fuzzing is failing with the proposed change