@lemire 2024-05-01
> fuzzing is failing with the proposed change

The change only affects ARM, and the fuzzer runs on x64.

I believe that oss-fuzz is busted, I have received many invalid bug reports and I have asked oss-fuzz to dismiss them. They are all of this type `MemorySanitizer: use-of-uninitialized-value` and they have to do with the standard library.